RAG/RAG MultiQueryretriever/requirements.txt
faiss-cpu
langchain-community
langchain-huggingface
python-dotenv
pypdf
cryptography
langchain_groq